Split blur kernel into multiple output loops.
Template the kernel on the number of outputs, and make each temporary
variable a fixed size array. Then invert the loop structure, so the
loop which computes the sum is the outer loop, and at each step it
loops over all outputs.

The idea was that this would ensure that each step had multiple
independent computations, thus increasing ILP. However, it's slower in
all cases. Most likely the loop overhead outweighs any ILP benefits.

However, I'm still committing it in order to preserve the code for
future reference.