Store sum separately for each output.
This is another, simpler attempt at improving ILP. By making the sums
independent and unrolling the loop, maybe instructions for each output
can be interleaved more effectively.

Performance seems identical to the simplest multi-output version which
keeps the sum in a single local variable, so there is no benefit to
the added complexity here. Committing to preserve the code for future
reference.