Direct convolution box blurs.
The results of this are surprising in two ways:

1. The horizontal blur is slightly faster than the vertical, even though
vertical ought to have better alignment of its memory access. Both
have a lot of redundant memory access to perhaps the cache performs
better for the horizontal pattern.

2. Both have similar speed to the sliding window version for widths up
to 20 or so. After that they slow down significantly.

Since both kernels are reasonably fast, we can use them together and
avoid the transposes. This can be up to 40% faster than the sliding
window method on small blurs (the crossover point is around width=15).