Implement direct Gaussian blurs.
This uses one vertical pass and one horizontal, computing the Gaussian
blur by direct convolution.

Calculating the weights is relatively expensive: this runs about 30%
slower than a single-pass of direct_blur_no_transpose, which has the
same memory access patterns. However, it's still faster than the
3-pass blur up to a radius of about 15.

The results are somewhat different from the repeated box blur (subtle
but visible with the naked eye if you know where to look), so we would
have pops if we switched from Gaussian to box as width changes.