Precompute weights for direct Gaussian blur.
Weights are computed with an iterated box blur so the results match
the baseline algorithm.

The overhead of looking up weights is about half that of computing
them in the kernel, which translates into a 15% speedup. This method
is faster than the sliding window for radius < 25, and it beats all
the other small blur optimizations across the board.

However, its results still differ from baseline at the edges. This is
because the iterated box blur gets slightly weird there (and arguably
wrong). We will need to adjust one of these implementations to prevent
popping when we switch.