Fix edges for precomputed blur.
The iterated box blur behaves differently at the edge of the
image (see comment in blur.cu for details). In the precomputed blur,
use different weights at the edges in order to match these results and
prevent popping.

In the course of this the simple box blur in compute-weights.cxx
started to get more complicated, to the point that it made more sense
to reuse the existing CUDA blur. So I renamed to compute-weights.cu
and rewrote most of the code.