Extract common code from horizontal/vertical blurs.
For both the direct box blur and precomputed blur, extract a helper
method which uses a 1D index and can perform either a horizontal or
vertical blur. These functions take an indexer lambda which adds the
missing coordinate and converts the 1D index to a pixel index.

This consolidates a lot of code that was nearly identical between the
horizontal and vertical cases.