Multiple blur passes in a single kernel.
The idea is that this might improve cache performance by revisiting a
line for subsequent passes right away, before it gets evicted by the
rest of the image. In practice, the benefit is minimal. In some
configurations it seems to give a speedup comparable to running
multiple threads per column, but since it's incompatible with that
optimization it doesn't improve on the previous best performance.