Trim the benchmark to reduce runtime.
This was previously taking about 10 hours on my machine which is
inconveniently long.
 - Remove all half-float tests
 - Remove the single_kernel_blur test
 - Always use 1 output per thread
 - Limit threads per column to 3 instead of 4
 - Run each test 20 times instead of 50

The removed configurations were wining in a small number of cases, but
usually the difference was trivial, or they were actually tied with
another configuration.

Also reordered the tests so the standard blur runs first, because
earlier tests have win in the case of ties. This way the benchmark
will only report a non-standard blur as the winner if it's actually
faster.