Proof of concept for staggered blur.
Compute three sliding window passes in a single loop, staggered so
that each pass finishes writing a given pixel before the subsequent
pass needs it.

This has significantly better cache locality and is >10% faster with
1920x1080 images and a large radius.

Temp image management is tricky here and this version sidesteps that
by having a lot of temps.

There is also has a bug causing some vertical streak
artifacts (probably an off-by-one causing it to read uninitialized
memory in a later pass).